+++ static/js/datatables-pipeline.js
@@ -424,6 +424,14 @@ $(document).ready(function() {
     if (newsigma_max.value) {
       qry_url = qry_url + "&max_new_high_sigma=" + newsigma_max.value;
     };
+    let source_selection = document.getElementById("sourceSelectionSelect");
+    let source_selection_type = document.getElementById("sourceSelectionTypeSelect");
+    if (source_selection.value) {
+      qry_url = qry_url + "&source_selection=" + source_selection.value;
+    };
+    if (source_selection_type.value) {
+      qry_url = qry_url + "&source_selection_type=" +source_selection_type.value;
+    };
     table.ajax.url(qry_url);
     table.ajax.reload();
   });
@@ -440,7 +448,7 @@ $(document).ready(function() {
       'RelationsMinSelect', 'RelationsMaxSelect', 'SelavyMinSelect', 'SelavyMaxSelect',
       'NewSigmaMinSelect', 'NewSigmaMaxSelect', 'NeighbourMinSelect', 'NeighbourMaxSelect',
       'compactnessMinSelect', 'compactnessMaxSelect', 'objectNameInput', 'MinSnrMinSelect',
-      'MinSnrMaxSelect', 'MaxSnrMinSelect', 'MaxSnrMaxSelect',
+      'MinSnrMaxSelect', 'MaxSnrMinSelect', 'MaxSnrMaxSelect', 'sourceSelectionSelect',
       ];
     var input;
     for (input of inputs) {